backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2==2.9.9
python-dotenv==1.0.1
Pillow==10.3.0
django-filter==24.1
